converter_service/requirements.txt
acres==0.5.0
aiofiles==24.1.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
etelemetry==0.3.1
filelock==3.18.0
fitz==0.0.1.dev2
frontend==0.0.3
grpcio==1.73.1
h11==0.16.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
looseversion==1.3.0
lxml==6.0.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
protobuf==6.31.1
prov==2.1.1
puremagic==1.29
pydot==4.0.1
PyMuPDF==1.26.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pyxnat==1.6.3
rdflib==7.1.4
requests==2.32.4
scipy==1.16.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
traits==7.0.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
